✨ [FEAT] 시리즈 수정 기능 및 리팩토링

+++ src/main/java/team/nine/booknutsbackend/service/SeriesService.java
@@ -11,9 +11,9 @@ import team.nine.booknutsbackend.dto.request.SeriesRequest;
 import team.nine.booknutsbackend.dto.response.BoardResponse;
 import team.nine.booknutsbackend.dto.response.SeriesResponse;
 import team.nine.booknutsbackend.exception.board.BoardNotFoundException;
-import team.nine.booknutsbackend.exception.user.NoAuthException;
 import team.nine.booknutsbackend.exception.series.SeriesDuplicateException;
 import team.nine.booknutsbackend.exception.series.SeriesNotFoundException;
+import team.nine.booknutsbackend.exception.user.NoAuthException;
 import team.nine.booknutsbackend.repository.BoardRepository;
 import team.nine.booknutsbackend.repository.SeriesBoardRepository;
 import team.nine.booknutsbackend.repository.SeriesRepository;
@@ -31,13 +31,20 @@ public class SeriesService {
     private final BoardRepository boardRepository;
     private final BoardService boardService;
 
-    //시리즈 목록 조회
+    //시리즈 조회
     @Transactional(readOnly = true)
-    public List<SeriesResponse> getSeriesList(User user) {
-        List<Series> stories = seriesRepository.findAllByOwner(user);
+    public Series getSeries(Long seriesId) {
+        return seriesRepository.findById(seriesId)
+                .orElseThrow(SeriesNotFoundException::new);
+    }
+
+    //특정 유저의 시리즈 목록 조회
+    @Transactional(readOnly = true)
+    public List<SeriesResponse> getSeriesList(User owner) {
+        List<Series> seriesList = seriesRepository.findAllByOwner(owner);
         List<SeriesResponse> seriesResponseList = new ArrayList<>();
 
-        for (Series series : stories) {
+        for (Series series : seriesList) {
             seriesResponseList.add(SeriesResponse.seriesResponse(series));
         }
 
@@ -61,11 +68,10 @@ public class SeriesService {
         return series;
     }
 
-    //특정 시리즈 조회
+    //특정 시리즈 내의 게시글 조회
     @Transactional(readOnly = true)
-    public List<BoardResponse> getSeries(Long seriesId, User user) {
-        Series series = seriesRepository.findById(seriesId)
-                .orElseThrow(SeriesNotFoundException::new);
+    public List<BoardResponse> getSeriesBoards(Long seriesId, User user) {
+        Series series = getSeries(seriesId);
         List<SeriesBoard> seriesBoards = seriesBoardRepository.findBySeries(series);
         List<BoardResponse> boardList = new ArrayList<>();
 
@@ -80,8 +86,9 @@ public class SeriesService {
     //시리즈 삭제
     @Transactional
     public void deleteSeries(Long seriesId, User user) {
-        Series series = seriesRepository.findBySeriesIdAndOwner(seriesId, user)
-                .orElseThrow(NoAuthException::new);
+        Series series = getSeries(seriesId);
+        if (series.getOwner() != user) throw new NoAuthException();
+
         List<SeriesBoard> seriesBoards = seriesBoardRepository.findBySeries(series);
 
         seriesBoardRepository.deleteAll(seriesBoards);
@@ -90,9 +97,10 @@ public class SeriesService {
 
     //시리즈에 게시글 추가
     @Transactional
-    public void addPostToSeries(Long seriesId, Long boardId) {
-        Series series = seriesRepository.findById(seriesId)
-                .orElseThrow(SeriesNotFoundException::new);
+    public void addPostToSeries(Long seriesId, Long boardId, User user) {
+        Series series = getSeries(seriesId);
+        if (series.getOwner() != user) throw new NoAuthException();
+
         Board board = boardRepository.findById(boardId)
                 .orElseThrow(BoardNotFoundException::new);
 
@@ -105,6 +113,19 @@ public class SeriesService {
         seriesBoard.setBoard(board);
         series.setOwner(series.getOwner());
         seriesBoardRepository.save(seriesBoard);
+    }
+
+    //시리즈 수정
+    @Transactional
+    public Series updateSeries(Long seriesId, SeriesRequest seriesRequest, User user) {
+        Series series = getSeries(seriesId);
+        if (series.getOwner() != user) throw new NoAuthException();
+
+        if (seriesRequest.getTitle() != null) series.setTitle(seriesRequest.getTitle());
+        if (seriesRequest.getContent() != null) series.setContent(seriesRequest.getContent());
+        if (seriesRequest.getImgUrl() != null) series.setImgUrl(seriesRequest.getImgUrl());
+
+        return seriesRepository.save(series);
     }
 
 }
